pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "CITE-seq-Count"
version = "1.6.0"
description = "A python package to map reads from CITE-seq or hashing data for single cell experiments"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-levenshtein",
    "scipy ==1.11.3",
    "pandas ==2.1.1",
    "pybktree ==1.1",
    "umi_tools",
    "multiprocess ==0.70.15",
    "typer ==0.9.0",
    "rich ==13.6.0",
    "loguru ==0.7.2",
    "tqdm ==4.66.1",
]

[[project.authors]]
name = "Roelli Patrick"
email = "patrick.roelli@gmail.com"

[project.urls]
repository = "https://github.com/Hoohm/CITE-seq-Count/"
homepage = "https://hoohm.github.io/CITE-seq-Count/"
documentation = "https://hoohm.github.io/CITE-seq-Count/"

[project.scripts]
CITE-seq-Count = "cite_seq_count.__main__:app"

[project.optional-dependencies]
dev = [
    "pytest ==7.4.2",
    "pytest-dependency ==0.5.1",
    "black ==23.9.1",
    "ruff ==0.0.292",
    "mypy ==1.6.0",
    "pandas-stubs ==2.1.1.230928",
    "types-regex ==2023.10.3.0",
    "types-python-dateutil ==2.8.19.14",
    "types-tqdm ==4.66.0.2",
    "pytest-profiling ==1.7.0",
]

[tool.black]
line-length = 120
target-version = ["py310"]
skip-string-normalization = true

[tool.ruff]
target-version = "py310"
line-length = 120
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "I",
    "I001",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "B027",
    "C901",
    "D100",
    "D103",
    "D205",
    "D212",
    "D415",
    "E203",
    "E501",
    "FBT003",
    "N802",
    "N803",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "S105",
    "S106",
    "S107",
]
unfixable = [
    "F401",
    "UP007",
]
exclude = [
    ".git",
    ".hg",
    "__pycache__",
    "_bin/*",
    "_build/*",
    "_ig_fbcode_wheel/*",
    "buck-out/*",
    "third-party-buck/*",
    "third-party2/*",
    "dist",
    ".venv",
    ".nox",
    ".mypy_cache",
    ".pytype",
    ".svn",
    "__pypackages__",
]
src = ["src"]

[tool.ruff.mccabe]
max-complexity = 50

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.isort]
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 80

[tool.hatch.build.targets.wheel]
packages = ["src/cite_seq_count"]
